Added middleware and integrated with creating events, bookings and cancelling them

+++ graphql/resolvers/bookingResolvers.js
@@ -1,32 +1,35 @@
 const Booking = require('../../models/Booking')
-const { getEvent, getUser } = require('../../utils')
+const { getEvent, getUser, getEvents } = require('../../utils')
 
 module.exports = {
-  async bookings() {
+  async bookings(args, { isAuthenticated }) {
     try {
+      if (!isAuthenticated) throw new Error('Incorrect Credentials')
       const bookings = await Booking.find({}).sort('-createdAt')
       return bookings.map(b => ({
         ...b._doc,
-        event: getEvent(b.event),
-        user: getUser(b.user)
+        event: getEvent.bind(this, b.event),
+        user: getUser.bind(this, b.user)
       }))
     } catch (err) {
       throw err
     }
   },
 
-  async bookEvent({ eventId }) {
+  async bookEvent({ eventId }, { isAuthenticated, user }) {
     try {
-      const event = getEvent(eventId)
+      if (!isAuthenticated) throw new Error('Incorrect Credentials')
+      const event = await getEvent(eventId)
+      if (!event) throw new Error('Event not found')
       const booking = new Booking({
         event: eventId,
-        user: '5dd92b7878d796931da7239a'
+        user: user._id
       })
       await booking.save()
       return {
         ...booking.toObject(),
         event,
-        user: getUser('5dd92b7878d796931da7239a')
+        user: { ...user.toObject(), events: getEvents.bind(this, user.events) }
       }
     } catch (err) {
       console.log(err)
@@ -34,8 +37,9 @@ module.exports = {
     }
   },
 
-  async cancelBooking({ bookingId }) {
+  async cancelBooking({ bookingId }, { isAuthenticated }) {
     try {
+      if (!isAuthenticated) throw new Error('Incorrect Credentials')
       const booking = await Booking.findByIdAndDelete(bookingId)
       if (!booking) throw new Error('Booking not found')
       const event = await getEvent(booking.event)
